Time: 0 ms (100.00%), Space: 8.1 MB (53.52%) - LeetHub

+++ 0008-string-to-integer-atoi/0008-string-to-integer-atoi.cpp
@@ -4,7 +4,7 @@ public:
     {
         int i=0;
         int sign=1;
-        long ans=0;
+        long long ans=0;
         while(i<s.length() && s[i]==' ')
             i++;
         if(s[i]=='-')
